back/requirements.txt
fastapi[standard]
python-jose[cryptography]